New_Student_Bot/requirements.txt
Flask==3.0.3
python-docx==1.1.2
numpy==2.1.1
